refactor: move stores to reducers
Install react-redux and move reducers in store into reducers folder.
