+++ README.md
@@ -64,7 +64,7 @@ All documentation for the SL-protocol can be found in the [SL-Wiki](http://wiki.
 
 ### Known Issues
 
-If testing fails to start. You might need to install [Facebooks Watchman](https://facebook.github.io/watchman/). [More here]https://github.com/Terreii/andromeda-viewer/blob/latest/README-create-react-app.md).
+If testing fails to start. You might need to install [Facebooks Watchman](https://facebook.github.io/watchman/). [More here](https://github.com/Terreii/andromeda-viewer/blob/latest/README-create-react-app.md).
 
 ## Disclaimer
 [Second Life(TM)](https://secondlife.com) is a product by [Linden Lab](http://www.lindenlab.com/). Linden Lab is not involved with this project!
