+++ src/reducers/friendsReducer.js
